Continue moving command logic to core.
